move types from server to common to avoid duplication
